requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
comtypes==1.1.10
    # via pywinauto
psutil==5.8.0
    # via -r requirements.in
pywin32==302
    # via pywinauto
pywinauto==0.6.8
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
six==1.16.0
    # via pywinauto
